% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query.R
\name{add.gquery.attractors}
\alias{add.gquery.attractors}
\title{Add target attractors to a Griffin query.}
\usage{
add.gquery.attractors(j.query, df.attr)
}
\arguments{
\item{df.attr}{dataframe with valid target attractors}

\item{query}{Griffin query}
}
\value{
query java query to run Griffin
}
\description{
Add a set of target steady-state attractors to a Griffin query.
The function takes a dataframe with the attractors 
where wach row corresponds to a steady state and 
each column corresponds to a node of the network
The valid values are:
   1: active
   0: inactive
   *: non-determined
}
\examples{
attr = data.frame(a=c(0,'*',0,1), 
                  b=c(0,1,0,'*'), 
                  c=c(0,0,1,0),
                  stringsAsFactors = F )
q = add.gquery.attractors(q, attr)
}
