% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query.R
\name{add.gquery.cycle}
\alias{add.gquery.cycle}
\title{Add target cycle to a Griffin query.}
\usage{
add.gquery.cycle(j.query, df.cycle)
}
\arguments{
\item{df.cycle}{dataframe with cycle}

\item{query}{Griffin query}
}
\value{
query java query to run Griffin
}
\description{
Add a target cyclic attractor to a Griffin query.
The function takes a dataframe with the states of the cycle 
where wach row corresponds to a state in order and 
each column corresponds to a node of the network
The valid values are:
   1: active
   0: inactive
   *: non-determined
}
\examples{
cycle = data.frame(a=c(0,0), b=c(0,1), c=c(1,0),
                  stringsAsFactors = F )
q = add.gquery.cycle(q, cycle)
}
