% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query.R
\name{add.gquery.prohibited.attractors}
\alias{add.gquery.prohibited.attractors}
\title{Add prohibited attractors to a Griffin query.}
\usage{
add.gquery.prohibited.attractors(j.query, df.prohibited.attr)
}
\arguments{
\item{df.prohibited.attr}{dataframe with valid prohibited attractors}

\item{query}{Griffin query}
}
\value{
query java query to run Griffin
}
\description{
Add a set of prohibited steady-state attractors to a Griffin query.
The function takes a dataframe with the prohibited attractors 
where wach row corresponds to a prohibited steady state and 
each column corresponds to a node of the network
The valid values are:
   1: active
   0: inactive
   *: non-determined
}
\examples{
attr.prohibited = data.frame(a=c('*'), b=c(1), c=c(1),
                  stringsAsFactors = F )
q = add.gquery.prohibited.attractors(q, attr.prohibited)
}
