% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query.R
\name{add.gquery.transition}
\alias{add.gquery.transition}
\title{Add a transition between two states to a Griffin query.}
\usage{
add.gquery.transition(j.query, df.transition)
}
\arguments{
\item{j.query}{Griffin query}

\item{df.transition}{dataframe with valid states in order}
}
\value{
query java query to run Griffin
}
\description{
Add a transition between two states to a Griffin query.
The function takes a dataframe with the steady states
where wach row corresponds to a steady state and 
each column corresponds to a node of the network
The valid values are:
   1: active
   0: inactive
   *: non-determined
}
\examples{
transition = data.frame(a=c(1,1), b=c(0,1), c=c(0,0),
                  stringsAsFactors = F )
q = add.gquery.transition(q, transition)
}
