% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQuery.R
\name{addGqueryAttractors}
\alias{addGqueryAttractors}
\title{Add target attractors to a Griffin query.}
\usage{
addGqueryAttractors(j.query, df.attr)
}
\arguments{
\item{j.query}{Griffin query}

\item{df.attr}{dataframe with valid target attractors}
}
\value{
query java query to run Griffin
}
\description{
Add a set of target steady-state attractors to a Griffin query.
The function takes a dataframe with the attractors 
where wach row corresponds to a steady state and 
each column corresponds to a node of the network
The valid values are:
   1: active
   0: inactive
   *: non-determined
}
\examples{
attr = data.frame(a=c(0,'*',0,1), 
                  b=c(0,1,0,'*'), 
                  c=c(0,0,1,0),
                  stringsAsFactors = F )
q = addGqueryAttractors(q, attr)
}
