% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQueryAttractors.R
\name{addGqueryCycle}
\alias{addGqueryCycle}
\title{Add target cycle to a Griffin query.}
\usage{
addGqueryCycle(j.query, df.cycle)
}
\arguments{
\item{j.query}{Griffin query}

\item{df.cycle}{dataframe with cycle}
}
\value{
query java query to run Griffin
}
\description{
Add a target cyclic attractor to a Griffin query.
The function takes a dataframe with the states of the cycle 
where wach row corresponds to a state in order and 
each column corresponds to a node of the network
The valid values are:
   1: active
   0: inactive
   *: non-determined
}
\examples{
cycle = data.frame(a=c(0,0), b=c(0,1), c=c(1,0),
                  stringsAsFactors = F )
q = addGqueryCycle(q, cycle)
}
