% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQuery.R
\name{addGqueryMutant}
\alias{addGqueryMutant}
\title{Add mutant with attractors to a Griffin query.}
\usage{
addGqueryMutant(j.query, mutant.nodes, mutant.values, df.mutant.attr)
}
\arguments{
\item{j.query}{Griffin query}

\item{mutant.nodes}{nodes whose value is fixed in the mutant}

\item{mutant.values}{values of the nodes in the mutant}

\item{df.mutant.attr}{dataframe with valid attractors}
}
\value{
query java query to run Griffin
}
\description{
Add a mutant and a set of target steady-state attractors to a Griffin query.
The function takes a vector of mutant nodes, a vector of its values
and a dataframe with the target attractors for the mutant
where wach row corresponds to a prohibited steady state and 
each column corresponds to a node of the network
The valid values are:
   1: active
   0: inactive
   *: non-determined
}
\examples{
# one node mutant
mutant.node = c('c')
mutant.value = c(1) 
mutant.attr = data.frame(a=c(0,1), b=c(0,0), c=c(1,1),
                         stringsAsFactors = F )
q = addGqueryMutant(q, mutant.node, mutant.value, mutant.attr)

# multiple node mutant
mutant.nodes = c('a','b')
mutant.values = c(0,0) 
mutant.attrs = data.frame(a=c(0), b=c(0), c=c('*'),
                         stringsAsFactors = F )
q = addGqueryMutant(q, mutant.nodes, mutant.values, mutant.attrs)
}
