% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQuery.R
\name{addGqueryTrapspace}
\alias{addGqueryTrapspace}
\title{Add a trapspace to a Griffin query.}
\usage{
addGqueryTrapspace(j.query, df.trapspace)
}
\arguments{
\item{j.query}{Griffin query}

\item{df.trapspace}{dataframe with valid trapspace}
}
\value{
query java query to run Griffin
}
\description{
Add a trapspace to a Griffin query.
The function takes a dataframe with the trapspace where the row corresponds to it 
each column corresponds to a node of the network
The valid values are:
   1: active
   0: inactive
   *: non-determined
}
\examples{
trap = data.frame(a=c(1), b=c('*'), c=c('*'),
                  stringsAsFactors = F )
q = addGqueryTrapspace(q, trap)
}
