% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQuery.R
\name{boolnetModelIterator}
\alias{boolnetModelIterator}
\title{Iterates over griffin controller to return string networks}
\usage{
boolnetModelIterator(controller, n = 1)
}
\arguments{
\item{controller}{griffin controller created by run.query}

\item{n}{number of networks to return}
}
\description{
Iterates over griffin controller to return string networks
}
\keyword{internal}
