% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query.R
\name{create.gquery.graph}
\alias{create.gquery.graph}
\title{Create a Griffin query.}
\usage{
create.gquery.graph(df.graph, nodes, n_source = "source",
  n_target = "target", n_reg = "type")
}
\arguments{
\item{df.graph}{dataframe with source and target nodes and the type of interactions}

\item{nodes}{vector with all node names}

\item{n_source}{name of column with source nodes, defaul='source'}

\item{n_target}{name of column with target nodes, defaul='target'}

\item{n_reg}{name of column with type of interaction, defaul='type'}
}
\value{
query java query to run Griffin
}
\description{
This function takes an interaction regulatory graph and creates a griffin query.
}
\details{
The graph must include: source, target and type of interaction
Valid types of interctions are:
   false:  Contradiction
   MA: Mandatory, ambiguous
   MPU (or +): Mandatory, positive, unambiguous
   MPPA: Mandatory, positive, possibly ambiguous
   MNU (or -): Mandatory, negative, unambiguous
   MNPA: Mandatory, negative, possibly ambiguous
   MUSU: Mandatory, unknown sign, unambiguous
   MUSPA:  Mandatory, unknown sign, possibly ambiguous
   NR: No regulation
   OA: Optional, ambiguous
   OPU:  Optional, positive, unambiguous
   OPPA: Optional, positive, possibly ambiguous
   ONU:  Optional, negative, unambiguous
   ONPA: Optional, negative, possibly ambiguous
   OUSU: Optional, unknown sign, unambiguous
   true: Tautology
}
\examples{
> genes = c('a','b','c')
> inter = data.frame(source=c('a','b','b','c','c'), 
                     target=c('b','b','c','b','c'), 
                     type=c('+','+','+','-','+'),
                     stringsAsFactors = F )
> create.gquery.graph(inter, genes)
}
