% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griffinToBoolnet.R
\name{dataframeToAttractor}
\alias{dataframeToAttractor}
\title{Convert a data frame with nodes displayed in Boolean format to a BoolNet attractor.}
\usage{
dataframeToAttractor(df, node.names, fixed.genes)
}
\arguments{
\item{df}{dataframe, see\code{\link{attractor2dataframe}} each column corresponds to the number of attractor, state, or node}

\item{node.names}{nodes of network}

\item{fixed.genes}{fixedGenes of network}
}
\value{
attr BoolNet attractor object
}
\description{
Convert a data frame with nodes displayed in Boolean format to a BoolNet attractor. First column is the attractor number, second is the number of state inside the attractor, the rest of the columns correspond to each node.
}
\examples{
> data("cellcycle")
> attr <- getAttractors(cellcycle)
> attr.df <- attractorToDataframe(attr)
> print(dataframe2attractor(attr.df, cellcycle$genes))

}
