% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{.onAttach}
\alias{.onAttach}
\title{TODO:
[ ] In documentation definition of r-graph
[ ] Automathicaly detect griffin path}
\usage{
.onAttach(lib, pkg, ...)
}
\description{
Future:
[ ] add.gquery.mutant.cycle <- function(nodes, values, df){future}
[ ] add.gquery.functions <- function(df) {future}
[ ] add.gquery.labels  <- function(df) {future} 
Initialize jvm
}
\keyword{internal}
