% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAttractors.R
\name{find.attractors}
\alias{find.attractors}
\title{Obtain the attractors and basins from a BoolNet network using symbolic methods.}
\usage{
find.attractors(net, return = c("DataFrame", "AttractorInfo"))
}
\arguments{
\item{net}{BoolNet BooleanNetwork}

\item{return}{"DataFrame"  returns a dataframe with the attractor in Boolean format
"AttractorInfo"  returns a BoolNet attractor object. 
where cyclic states may use more than one row. If requiered the 
method will return the basinSize and basinFormula as columns in the 
attr$attractors object. For cyclic attractor it will add the properties 
in all rows of the states of the attractor.}
}
\value{
If "DataFrame" it will return a dataframe wherethe first column corresponds to the number of the atrractor and the second to the state in the case of cyclic atractors the attractor will take multiple states. The rest of the columns correspond to the value if he nodes, tha basinSize and basiFormula
      If "AtractorInfo" return a BoolNet attractor object where basinSize and basinFormula are in attactor$attractors
}
\description{
This function takes a valid BoolNet BooleanNetwork.
and returns the attractors, basin size and basin formula of the network using a binary decision tree based method.
To avoid problems use fully parantherized formulas. Griffin uses a left-right formula reader that may cause problems.
}
\examples{
data("cellcycle")
attr = find.attractors( cellcycle )
print( attr )

}
\seealso{
Compare and contrast with \code{\link{BoolNet:getAttractors}} SAT based options.
}
