% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolnet2griffin.R
\name{get.interaction.sign}
\alias{get.interaction.sign}
\title{Determine the sign of the interaction between two elements of a network}
\usage{
get.interaction.sign(source, target, net)
}
\arguments{
\item{source}{source node name}

\item{target}{target node name}

\item{net}{BoolNet function}
}
\value{
interaction str interaction type between source and target
}
\description{
Returns a dataframe with the following interaction types:
    '+'  : Mandatory, positive, unambiguous
    '-'  : Mandatory, negatice, unambiguous
    'MA' : Mandatory, ambiguous
    'NR' : No regulation
    NULL : No interaction
}
\keyword{internal}
