% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBasins.R
\name{getBasins}
\alias{getBasins}
\title{Obtain the attractors and basins from a BoolNet network using symbolic methods.}
\usage{
getBasins(net, return = c("DataFrame", "AttractorInfo"))
}
\arguments{
\item{net}{BoolNet BooleanNetwork}

\item{return}{"DataFrame"  returns a dataframe with the attractor in Boolean
format where cyclic states may use more than one row. The 
basinSize and basinFormula are added as columns
"AttractorInfo"  returns a BoolNet attractor object. The 
basinSize and basinFormula are added to attractor$attractor.}
}
\value{
If "DataFrame" it will return a dataframe wherethe first column corresponds to the number of the atrractor and the second to the state in the case of cyclic atractors the attractor will take multiple states. The rest of the columns correspond to the value if he nodes, tha basinSize and basiFormula.
      If "AtractorInfo" return a BoolNet attractor object where basinSize and basinFormula are in attactor$attractors
}
\description{
This function takes a valid BoolNet BooleanNetwork.
and returns the attractors and, if requiered basin size and basin formula of the network using a binary decision tree based method.
The process may be memory intensive for large networks. To avoid 
edge cases use fully parantherized formulas.
}
\examples{
data("cellcycle")
attr = getBasins( cellcycle )
print( attr )

}
\seealso{
Compare and contrast with \code{\link{BoolNet:getAttractors}} SAT based options.
}
