% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolnetToGriffin.R
\name{getNetTopology}
\alias{getNetTopology}
\title{Returns the topology with signs of a network}
\usage{
getNetTopology(net)
}
\arguments{
\item{net}{BoolNet function}
}
\value{
df.graph dataframe with source and target nodes and the type of interactions
}
\description{
Takes a BoolNet network
Returns a DataFrame of the form: Source, Target, Interaction
Interactions types can be:
    '+'  : Mandatory, positive, unambiguous
    '-'  : Mandatory, negatice, unambiguous
    'MA' : Mandatory, ambiguous
    'NR' : Non-functional regulation
}
