% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_query.R
\name{griffin.model.iterator}
\alias{griffin.model.iterator}
\title{Iterates over griffin controller to return string networks}
\usage{
griffin.model.iterator(controller, n = 1)
}
\arguments{
\item{controller}{griffin controller created by run.query}

\item{n}{number of networks to return}
}
\description{
Iterates over griffin controller to return string networks
}
\keyword{internal}
