% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{init.griffin}
\alias{init.griffin}
\title{Initialize the Java Virtual Machine (JVM) and create a connection with griffin}
\usage{
init.griffin(jvm.param = NULL, force.init = TRUE)
}
\arguments{
\item{jvm.param}{string with parameters to initialize the jvm.}

\item{force.init}{If set to TRUE JVM is re-initialized even if it is already running.}
}
\description{
Initialize the Java Virtual Machine (JVM) and create a connection with griffin.
By default it takes the parameters from the "java/jvm-param.R" file. To change the defaults modify this file in your rgriffin folder
}
