% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gquery_class.R
\name{print.jobjRef}
\alias{print.jobjRef}
\title{setClass("gquery",contains = "jobjRef")}
\usage{
\method{print}{jobjRef}(j.query)
}
\description{
#' print query
#' @export
print.gquery <- function(j.query) {
  text = .jcall(j.query,returnSig = "S","asString",evalString = TRUE)
  text = gsub('^.|.$', '', text)
  text = gsub("Gene Regulation Specification. ", "Gene Regulation Specification; ", text)
  text = gsub(" = true", "=1", text)
  text = gsub(" = false", "=0", text)
  text = gsub(";", "\n", text)
  cat(text)
}
print query
}
\keyword{internal}
