% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griffin2boolnet.R
\name{string2boolnet}
\alias{string2boolnet}
\title{Create a BoolNet net from a string}
\usage{
string2boolnet(griffin.str.net, bodySeparator = ",",
  lowercaseGenes = FALSE, symbolic = FALSE)
}
\description{
Create a BoolNet net from a string
}
\keyword{internal}
