% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griffinToBoolnet.R
\name{stringToBoolnet}
\alias{stringToBoolnet}
\title{Create a BoolNet net from a string}
\usage{
stringToBoolnet(griffin.str.net, bodySeparator = ",",
  lowercaseGenes = FALSE, symbolic = FALSE)
}
\description{
Create a BoolNet net from a string
}
\keyword{internal}
