% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_griffin.R
\name{validate.states}
\alias{validate.states}
\title{Check if the states are of the correct size, order, characters and type.}
\usage{
validate.states(df.attr, nodes)
}
\description{
Check if the states are of the correct size, order, characters and type.
}
\keyword{internal}
