% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperGriffin.R
\name{validateStates}
\alias{validateStates}
\title{Check if the states are of the correct size, order, characters and type.}
\usage{
validateStates(df.attr, nodes)
}
\description{
Check if the states are of the correct size, order, characters and type.
}
\keyword{internal}
